+++ app/src/main/java/com/dming/testplayer/OnProgressListener.java
@@ -2,6 +2,6 @@ package com.dming.testplayer;
 
 public interface OnProgressListener {
 
-    void onProgress(long curTime,long totalTime);
+    void onProgress(int curTime,int totalTime);
 
 }
